UI Updates and Fixes
