+++ src/components/common/WebVitals.tsx
@@ -2,7 +2,7 @@
 
 import { useReportWebVitals } from "next/web-vitals";
 
-export function WebVitals() {
+export default function WebVitals() {
   useReportWebVitals((metric) => {
     console.log(metric);
   });
